api/table_reservation/requirements.txt
djangorestframework==3.16.0
django==5.2
psycopg2-binary==2.9.10